requirements.txt
pygatt
pycryptodome